requirements.txt
eth-brownie
brownie-token-tester
